more vision, started adding log methods
