+++ src/main/java/frc/robot/subsystems/NewVision/NewVisionSubsystem.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.NewVision;
+package frc.robot.subsystems.NewVision;
 
 import java.io.IOException;
 import java.util.Optional;
